corrections in distance calculation

+++ database/migrations/2020_04_26_145659_create_data_table.php
@@ -15,8 +15,8 @@ class CreateDataTable extends Migration
     {
         Schema::create('data', function (Blueprint $table) {
             $table->id();
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->double('latitude');
+            $table->double('longitude');
             $table->integer('distance');
             $table->string('location');
             $table->unsignedBigInteger('user_id');
